Improved accuracy by using chains of simplifyable calculations

+++ app/src/main/java/app/gaborbiro/freelancecalculator/persistence/domain/Store.kt
@@ -7,13 +7,13 @@ import java.math.BigDecimal
 
 interface Store {
 
-    var feePerHour: Flow<BigDecimal?>
+    var feePerHour: Flow<Double?>
 
-    var hoursPerWeek: Flow<BigDecimal?>
+    var hoursPerWeek: Flow<Double?>
 
-    var daysPerWeek: Flow<BigDecimal?>
+    var daysPerWeek: Flow<Double?>
 
-    var fee: Flow<BigDecimal?>
+    var fee: Flow<Double?>
 
     var fromCurrency: Flow<String?>
 
@@ -23,18 +23,18 @@ interface Store {
 
     companion object {
         fun dummyImplementation(
-            feePerHour: Int = 80,
-            hoursPerWeek: Int = 30,
-            daysPerWeek: Int = 5,
-            fee: Int = 8,
+            feePerHour: Double = 80.0,
+            hoursPerWeek: Double = 30.0,
+            daysPerWeek: Double = 5.0,
+            fee: Double = 8.0,
             fromCurrency: String = "USD",
             toCurrency: String = "GBP",
             sectionExpanded: Map<String, Boolean> = emptyMap(),
         ) = object : Store {
-            override var feePerHour: Flow<BigDecimal?> = flowOf(BigDecimal(feePerHour))
-            override var hoursPerWeek: Flow<BigDecimal?> = flowOf(BigDecimal(hoursPerWeek))
-            override var daysPerWeek: Flow<BigDecimal?> = flowOf(BigDecimal(daysPerWeek))
-            override var fee: Flow<BigDecimal?> = flowOf(BigDecimal(fee))
+            override var feePerHour: Flow<Double?> = flowOf(feePerHour)
+            override var hoursPerWeek: Flow<Double?> = flowOf(hoursPerWeek)
+            override var daysPerWeek: Flow<Double?> = flowOf(daysPerWeek)
+            override var fee: Flow<Double?> = flowOf(fee)
             override var fromCurrency: Flow<String?> = flowOf(fromCurrency)
             override var toCurrency: Flow<String?> = flowOf(toCurrency)
             override fun sectionExpander(): TypedSubStore<Boolean> =
